Moving fetch to Hooks
