+++ frontEnd/src/hooks/useProducts.js
@@ -1,0 +1,28 @@
+import { useState, useEffect } from 'react';
+import { getProducts } from '../services/productService';
+
+export const useProducts = () => {
+    const [products, setProducts] = useState([]);
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(null);
+
+    useEffect(() => {
+        const fetchProducts = async () => {
+            try {
+                const data = await getProducts();
+                setProducts(data);
+            } catch (error) {
+                setError(error);
+            } finally {
+                setLoading(false);
+            }
+        };
+        fetchProducts();
+    }, []);
+
+    return {
+        products,
+        loading,
+        error
+    };
+};
